refactor(host): split odt mcp module
Break the host ODT MCP facade into focused mapping, types, task resolution, and service files so the public bridge logic stays reviewable without changing behavior.
